# src/main/resources/static/openapi.yml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bankrest API"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
    "/api/v1/users/{id}": {
      "get": {
        "tags": [
          "user-rest-controller"
        ],
        "operationId": "findById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserReadDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "user-rest-controller"
        ],
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateEditDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserReadDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "user-rest-controller"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/transfers/{id}": {
      "get": {
        "tags": [
          "transfer-rest-controller"
        ],
        "operationId": "findById_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransferReadDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "transfer-rest-controller"
        ],
        "operationId": "update_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferCreateEditDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransferReadDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "transfer-rest-controller"
        ],
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/cards/{id}": {
      "get": {
        "tags": [
          "card-rest-controller"
        ],
        "operationId": "findById_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardReadDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "card-rest-controller"
        ],
        "operationId": "update_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardCreateEditDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardReadDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "card-rest-controller"
        ],
        "operationId": "delete_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/cards/{id}/blocked": {
      "put": {
        "tags": [
          "card-rest-controller"
        ],
        "operationId": "blockingCard",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardReadDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users": {
      "get": {
        "tags": [
          "user-rest-controller"
        ],
        "operationId": "findAll",
        "parameters": [
          {
            "name": "userFilter",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UserFilter"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseUserReadDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "user-rest-controller"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateEditDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserReadDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{userId}/transfers": {
      "get": {
        "tags": [
          "transfer-rest-controller"
        ],
        "operationId": "findAllByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseTransferReadDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "transfer-rest-controller"
        ],
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferCreateEditDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransferReadDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cards": {
      "get": {
        "tags": [
          "card-rest-controller"
        ],
        "operationId": "findAll_1",
        "parameters": [
          {
            "name": "cardFilter",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CardFilter"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseCardReadDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "card-rest-controller"
        ],
        "operationId": "create_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardCreateEditDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardReadDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/authentication": {
      "post": {
        "tags": [
          "authentication-controller"
        ],
        "operationId": "getToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccessToken"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{userId}/cards": {
      "get": {
        "tags": [
          "card-rest-controller"
        ],
        "operationId": "findAllByUserId_1",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "cardFilter",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CardFilter"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseCardReadDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transfers": {
      "get": {
        "tags": [
          "transfer-rest-controller"
        ],
        "operationId": "findAll_2",
        "parameters": [
          {
            "name": "transferFilter",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TransferFilter"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseTransferReadDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cards/{id}/balance": {
      "get": {
        "tags": [
          "card-rest-controller"
        ],
        "operationId": "getBalance",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "number"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserCreateEditDto": {
        "type": "object",
        "properties": {
          "firstname": {
            "type": "string",
            "maxLength": 64,
            "minLength": 2
          },
          "lastname": {
            "type": "string",
            "maxLength": 64,
            "minLength": 2
          },
          "phoneNumber": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ]
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "passportNumber": {
            "type": "string",
            "maxLength": 9,
            "minLength": 9
          },
          "password": {
            "type": "string",
            "maxLength": 64,
            "minLength": 3
          }
        },
        "required": [
          "birthDate",
          "role"
        ]
      },
      "UserReadDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ]
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "passportNumber": {
            "type": "string"
          }
        }
      },
      "TransferCreateEditDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "cardFrom": {
            "type": "string",
            "maxLength": 16,
            "minLength": 16
          },
          "cardTo": {
            "type": "string",
            "maxLength": 16,
            "minLength": 16
          },
          "transferDate": {
            "type": "string",
            "format": "date"
          },
          "sum": {
            "type": "number"
          }
        },
        "required": [
          "sum",
          "userId"
        ]
      },
      "TransferReadDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userReadDto": {
            "$ref": "#/components/schemas/UserReadDto"
          },
          "cardFrom": {
            "type": "string"
          },
          "cardTo": {
            "type": "string"
          },
          "transferDate": {
            "type": "string",
            "format": "date"
          },
          "sum": {
            "type": "number"
          }
        }
      },
      "CardCreateEditDto": {
        "type": "object",
        "properties": {
          "number": {
            "type": "string",
            "maxLength": 16,
            "minLength": 16
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "expirationDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "BLOCKED",
              "EXPIRED"
            ]
          },
          "balance": {
            "type": "number"
          }
        },
        "required": [
          "balance",
          "expirationDate",
          "status",
          "userId"
        ]
      },
      "CardReadDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "string"
          },
          "userReadDto": {
            "$ref": "#/components/schemas/UserReadDto"
          },
          "expirationDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "BLOCKED",
              "EXPIRED"
            ]
          },
          "balance": {
            "type": "number"
          }
        }
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "maxLength": 64,
            "minLength": 3
          }
        }
      },
      "AccessToken": {
        "type": "object",
        "properties": {
          "IdToken": {
            "type": "string"
          }
        }
      },
      "UserFilter": {
        "type": "object",
        "properties": {
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "passportNumber": {
            "type": "string"
          }
        }
      },
      "Metadata": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElement": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "PageResponseUserReadDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserReadDto"
            }
          },
          "metadata": {
            "$ref": "#/components/schemas/Metadata"
          }
        }
      },
      "PageResponseTransferReadDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransferReadDto"
            }
          },
          "metadata": {
            "$ref": "#/components/schemas/Metadata"
          }
        }
      },
      "CardFilter": {
        "type": "object",
        "properties": {
          "number": {
            "type": "string"
          },
          "expirationDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "PageResponseCardReadDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CardReadDto"
            }
          },
          "metadata": {
            "$ref": "#/components/schemas/Metadata"
          }
        }
      },
      "TransferFilter": {
        "type": "object",
        "properties": {
          "cardFrom": {
            "type": "string"
          },
          "cardTo": {
            "type": "string"
          },
          "transferDate": {
            "type": "string",
            "format": "date"
          }
        }
      }
    }
  }
}